serviceTest

+++ src/main/java/kenigsberg/transit/VehicleLocation.java
@@ -1,6 +1,6 @@
 package kenigsberg.transit;
 
 public class VehicleLocation {
-    long Longitude;
-    long Latitude;
+    float Longitude;
+    float Latitude;
 }
